requirements.txt
wget==3.2
timm==0.4.12
tqdm==4.61.2
numpy==1.19.5
Pillow==8.4.0
PyYAML==5.4.1
scikit_learn==0.24.2
